Added end to end API testing
