+++ application/app/src/test/java/uk/nhs/digital/uec/api/integration/fuzzysearch/FuzzySearchGeneralErrorsTest.java
@@ -1,0 +1,68 @@
+package uk.nhs.digital.uec.api.integration.fuzzysearch;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.client.TestRestTemplate;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.test.context.ActiveProfiles;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@ActiveProfiles({"prod"})
+/**
+ * Test class which passes requests through the Fuzzy Search endpoint and asserts desired API
+ * behavior. Only the model layer will be mocked here.
+ */
+public class FuzzySearchGeneralErrorsTest {
+
+  private static String endpointUrl =
+      "http://localhost:9095/dosapi/dosservices/v0.0.1/services/byfuzzysearch";
+
+  TestRestTemplate restTemplate = new TestRestTemplate();
+  HttpHeaders headers = new HttpHeaders();
+
+  /** Given a Post call on the endpoint, ensure a 405 error is returned. */
+  @Test
+  public void postCallOnEndpoint() throws Exception {
+    HttpEntity<String> request = new HttpEntity<String>(null, headers);
+    ResponseEntity<String> response =
+        restTemplate.exchange(endpointUrl, HttpMethod.POST, request, String.class);
+
+    assertTrue(response.getStatusCode() == HttpStatus.METHOD_NOT_ALLOWED);
+  }
+
+  /** Given a Patch call on the endpoint, ensure a 405 error is returned. */
+  @Test
+  public void patchCallOnEndpoint() throws Exception {
+    HttpEntity<String> request = new HttpEntity<String>(null, headers);
+    ResponseEntity<String> response =
+        restTemplate.exchange(endpointUrl, HttpMethod.PATCH, request, String.class);
+
+    assertTrue(response.getStatusCode() == HttpStatus.METHOD_NOT_ALLOWED);
+  }
+
+  /** Given a Put call on the endpoint, ensure a 405 error is returned. */
+  @Test
+  public void putCallOnEndpoint() throws Exception {
+    HttpEntity<String> request = new HttpEntity<String>(null, headers);
+    ResponseEntity<String> response =
+        restTemplate.exchange(endpointUrl, HttpMethod.PUT, request, String.class);
+
+    assertTrue(response.getStatusCode() == HttpStatus.METHOD_NOT_ALLOWED);
+  }
+
+  /** Given a Delete call on the endpoint, ensure a 405 error is returned. */
+  @Test
+  public void deleteCallOnEndpoint() throws Exception {
+    HttpEntity<String> request = new HttpEntity<String>(null, headers);
+    ResponseEntity<String> response =
+        restTemplate.exchange(endpointUrl, HttpMethod.DELETE, request, String.class);
+
+    assertTrue(response.getStatusCode() == HttpStatus.METHOD_NOT_ALLOWED);
+  }
+}
